[Response Ops][Rules] Rule Model Version Forward Compat (#193233)
## Summary

This PR implements the `forwardCompatibility` field for the rule model
version in preparation for the flapping PR
(https://github.com/elastic/kibana/pull/190019). We're doing this so we
have the point to revert back to if we need to where the
`forwardCompatibility` is defined.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit bb1899653f6fc077d443a600bce6c6f2fc775964)