Fix memory leak in task manager task runner (#193612)
In this PR, I'm fixing a memory leak that was introduced in
https://github.com/elastic/kibana/pull/190093 where every task runner
class object wouldn't free up in memory because it subscribed to the
`pollIntervalConfiguration$` observable. To fix this, I moved the
observable up a class into `TaskPollingLifecycle` which only gets
created once on plugin start and then pass down the pollInterval value
via a function call the task runner class can call.