Auto-focus ES|QL editor on mount
Currently, when the page loads, in order to start interacting with the ES|QL editor using keyboard navigation, you must:

* Tab to the "Skip to main content" link
* Follow that link
* Tab through the tab and toolbar elements until you get to the editor
* Press <end> or <ctrl-e> or down-down-down-down, depending on what's in the editor text box
* Type my query

This change auto-focuses the editor and moves the cursor to the end when the editor is mounted.

The [ARIA Authoring Practices Guide (APG) for Developing a Keyboard Interface](https://www.w3.org/WAI/ARIA/apg/practices/keyboard-interface/) says:

> * Do not set initial focus when the page loads except in cases where:
> ** The page offers a single, primary function that nearly all users employ immediately after page load.
> ** Any given user is likely to use the page often.

When using ES|QL in discover you almost always want to initially focus on the query, look at the results, and further refine the query.