[Advanced Settings] Fix search does not work for terms with ":" (#193506)
## Summary

Fixes #192694

When terms with `:`, it will be transformed to a `FieldClause`.

So, I'm try to set it back to a `TermClause` if the field is not in
`FieldDefinition`.

Example: `-defaultValue:false doc_table:high Dashboard`

![image](https://github.com/user-attachments/assets/e9519cfe-c64d-4bd4-b928-d6ffd9bfaa8b)

---------

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
Co-authored-by: Elena Stoeva <59341489+ElenaStoeva@users.noreply.github.com>
(cherry picked from commit b8398c7b684d9fb4745c0d9d8c5a9c60a5eddc48)