[Alert details page] Use related alerts logic only for the custom threshold rule (#193957)
Closes #193943

### Summary

This PR only loads related alerts for the custom threshold rule and only
when the related rule has tags or group by fields.

For now, we don't show this tab if there is no tag or group and there is
a follow-up ticket for discussing the empty state for the related
alerts.

#### Follow-up work
- https://github.com/elastic/kibana/issues/193952
- https://github.com/elastic/kibana/issues/193942

(cherry picked from commit 387afb782e44a0bf7de5549751e52fc7619c328a)