[workflows] Add condition to if backport:skip event
To only run when the label is removed and a backport:* label exists.

The on-merge workflow has historically supported version labels without
the existence of backport labels when a pull request is closed.
This isn't an expected use case now that we've enforced the label check,
but it still is possible to trigger.

Testing

1) Merge pull request
2) Add v8.16.0 label
3) Remove backport:skip label (current trigger 1)
4) Add backport:prev-minor label (current trigger 2)

Should trigger the action once.

This PR is split into two commits, the first with the condition and the
second with the changes run through a formatter.