[ES|QL] Don't store duplicate queries even of different formatting (#194267)
## Summary

One of the things that bother me in the history component is that you
can have different spaces in the queries but they query is actually the
same. These appear 2 times which I find very annoying.

This PR identifies if the query is the same, regardless of the
formatting, and if yes it will display it only once (it actually updates
the cache old item with the new one)

### Checklist
- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit 21a8bf8c20e4324312ce80d599245435d299d958)