[ML] Fix file upload when pipline is used (#194273)
Fixes bug introduced in https://github.com/elastic/kibana/pull/193744

If a file is uploaded with an ingest pipeline, the pipeline is created,
but not used when uploading the data.

Adds tests to check that the data exists in the index and the expected
fields exist and are populated.

(cherry picked from commit 6fd19137fe1cb0293ce84f3bba8f8ae7fedd98c9)