[Security Solution][Serverless] - Improve security solution performance (#194241)
## Summary

The goal of this PR is to improve the default performance of many of our
security solution views.

1. Upon scale testing, it was observed that the default events histogram
aggregation was a source of application slowness, so to improve the
performance of the default security experience, we've made the default
breakdown for the events histogram `No Breakdown` similar to what is
seen in the default discover histogram experience.

2. After looking through some telemetry, it was observed that the field
list query run in the background for timeline can also take a
significant amount of time based on the user's field count, so we now
only run that query after timeline has been opened.

### Demos
#### 1. By default the events visualizations on the overview and explore
events pages will not have an aggregation. The user will have to
manually select the breakdown they desire:
https://github.com/elastic/kibana/commit/d354d27962ebbd6d5fda19e912ec344ffe8a6c75

https://github.com/user-attachments/assets/a6d6987b-73fc-4735-9c37-973917c2fa2d

#### 2. Timeline fields list will only load after the first interaction
with timeline:
https://github.com/elastic/kibana/commit/ad557260d8f9c5dd0810a5a6aa51e5de0430000f

**Before:**

https://github.com/user-attachments/assets/0ad2e903-ac15-4daa-925b-da8ad05e80dd

**After:**

https://github.com/user-attachments/assets/27d5d3d5-02c8-49b5-b699-239ebc36b16c
(cherry picked from commit e45d97b26c6d0e0798d620ad0b097cad9009c179)