Make importRules functions more pure
This does away with the receiving, modifying, and returning of the
`rulesResponseAcc` parameter from both the new and old import paths.

This logic originally existed in the `import_rules_utils` interface, and
was cargo-culted to the new code path(s) incorrectly by me. It's not
necessary, so we simply do away with the complexity here.