[EDR Workflows] Fix agent count for single agent policies (#194294)
https://github.com/user-attachments/assets/2b64c1e0-0e6d-4ef5-952d-e4364b4403c4



The PR #193705 introduced an issue when counting active agents for
integration policies with only one agent policy assigned. In such cases,
`query.policyIds` was treated as a single string instead of an array of
strings (as expected with multiple agent policy ids like
`/?policyIds=x&policyIds=y`). This PR resolves the issue by ensuring
consistent handling of policyIds, regardless of the number of associated
agent policies.