Prevent unnecessary looping during import calculations
This simplifies the calculation of rule_source in cases where an asset
is found matching the `rule_id` but not the `version`. The logic is the
same, but we now use a Set() of `rule_id`s and perform a O(1) check with
`Set.has()` instead of performing `Array.includes` (O(n)) for each rule.

Since `calculateRuleSourceForImport` was previously performing this same
logic, and the Importer is the only caller, it's been updated to instead
receive a boolean representing the result of that calculation.