[Inventory] Use observability:entitycentricexperience to enable + side nav adjustments (#193533)
## Summary

Closes https://github.com/elastic/kibana/issues/192323 and
https://github.com/elastic/kibana/issues/193393.

### Plugin enablement
The enablement of the new Inventory plugin was previously managed via
the `xpack.inventory.enabled` setting in kibana.yml. With this PR, we
transition to enabling it through the
`observability:entitycentricexperience` advanced setting.


### Position in side nav
Additionally, this PR includes adjustments to the placement of the new
Inventory page within the side navigation and the addition of the "Tech
Preview" badge in the navigation where it was supported.

|Navigation|Screenshot|
|-|-|
|Stateful classic (first nav)|![Screenshot 2024-09-25 at 11 25
46](https://github.com/user-attachments/assets/54e92632-9931-4f5b-8648-8a32afcd0caf)|
|Stateful classic (O11y)|![Screenshot 2024-09-25 at 11 25
56](https://github.com/user-attachments/assets/ae644eb2-2559-4be0-8659-ef709399a871)|
|Stateful new O11y|![Screenshot 2024-09-25 at 11 27
52](https://github.com/user-attachments/assets/c404232c-2630-4220-9bb8-315b0f0fc55f)|
|Serverless|![Screenshot 2024-09-25 at 11 23
07](https://github.com/user-attachments/assets/7f7e6c79-7f0f-4f7d-8fed-260d73ad4d3f)|


### Global search
A modification was made to the global search to prevent a duplicate
result from appearing during searches by removing the deeplinks when
registering the plugin. The UI changes are shown below.

>[!WARNING]
If this modification causes any other issues, please let me know in a
comment.

|With deeplink (before)|Without deeplink (after)|
|-|-|

|![global_search_with_deeplink](https://github.com/user-attachments/assets/3ef5641c-3add-4892-af2f-46f72ff1722e)|![global_search_without_deeplink](https://github.com/user-attachments/assets/8ad92f27-13a2-4b85-8958-588e597f94d9)|

### Services rename to Service Inventory
Another needed change was to rename the nav item Services to Service
Inventory

>[!NOTE] 
Screenshots were taken before the navigation sorting was agreed with
product, so the "Inventory" item does not appear in the right position.

|Navigation|Screenshot|
|-|-|
|Stateful classic
(O11y)|![services_stateful_classic](https://github.com/user-attachments/assets/25ed1294-40fe-47be-9319-fe294acce8f8)|
|Stateful new
O11y|![services_stateful_new](https://github.com/user-attachments/assets/4ae9495c-56bc-4327-b2f6-b6ac9f68bd21)|

|Serverless|![services_serverless](https://github.com/user-attachments/assets/81ff51f4-3b0e-41fd-b7c6-79e8aa44e575)|

>[!NOTE] 
The rename of infra Inventory to Infrastructure Inventory is being
tracked separately in https://github.com/elastic/kibana/issues/192324.


### Tech preview badge in page
Lastly, the "Technical Preview" badge has been added to the page title.

>[!NOTE] 
The screenshot was taken before the navigation sorting was agreed with
product, so the "Inventory" item does not appear in the right position.



![inventory_title_badge](https://github.com/user-attachments/assets/172c5b8e-258e-4127-b792-40910e9dede0)


### Plugin registration in server
To make the Inventory available in the spaces and privileges
configuration the plugin has been registered in the features service.

>[!WARNING]
I lack detailed context in this area, so I’ve added default empty
configurations for most attributes. If any adjustments or improvements
are needed, please let me know.

|Space config|Privileges|
|-|-|
|![Screenshot 2024-09-23 at 10 57
35](https://github.com/user-attachments/assets/1c3a45b6-fd67-4706-8fc4-ee42a9206787)|![Screenshot
2024-09-23 at 10 57
14](https://github.com/user-attachments/assets/a852c15d-133c-459d-be77-2d73e662ffdb)|

---------

Co-authored-by: Kate Patticha <aikaterini.patticha@elastic.co>
Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>