[React@18] Upgrade @types to React 18 (#194144)
## Summary

Part of https://github.com/elastic/kibana/issues/138222

This PR finilizes the breaking type fixes that are needed for upgrade to
React@18.
Most of the remaining issues are muted with "@ts-expect-error" are
tricky or could be runtime bugs that need to be looked at.

**Since the types changes are backward compatible (except the new APIs)
we can upgrade to @types/react@18 now so that we "save" the progress and
all the code in Kibana from now on is written in compatbile for react@18
way from types perspective.**

(cherry picked from commit bdd57b64fe19683c481f673d6f1571f91379e730)