Fix code scan alert
To fix the problem, we need to validate the type parameter before using it to
access a property in the Watch.getWatchTypes() object. We can achieve this by
checking if the type exists in the Watch.getWatchTypes() object and if it is a
function. If the type is invalid, we should handle the error appropriately.

1) Modify the code in the useEffect hook to validate the type parameter.
2) Ensure that the type corresponds to a valid key in the Watch.getWatchTypes() object
and that the value is a function.
3) If the type is invalid, handle the error by dispatching an appropriate error action.