[EDR Workflows][Serverless] Gate custom note (#193171)
This PR implements tier-based gating for custom notification messages in
Protections. Only users on the Endpoint Complete tier will have the
ability to modify these messages, while users on the Endpoint Essentials
tier will no longer have this capability. If a user on the Essentials
tier had made any changes to custom notifications before this update,
those messages will be reset to the default ones.

The changes are applied in three areas:
1. UI - An upsell banner is displayed for Essentials users.
2. API - We now prevent API calls that attempt to set or modify custom
notification messages for Essentials users.
3. Policy Watcher - Upon Kibana startup (e.g., after a downgrade), we
validate all policies for tier compliance. If a policy contains a custom
notification message and the user is on the Essentials tier, the message
will be reset to the default.

![Screenshot 2024-09-20 at 14 32
52](https://github.com/user-attachments/assets/75739338-e32b-47da-934e-9948f44099ae)
![Screenshot 2024-09-20 at 14 33
21](https://github.com/user-attachments/assets/1af081eb-f75f-4c9d-8f01-df9a01f8f2b2)
![Screenshot 2024-09-20 at 14 33
40](https://github.com/user-attachments/assets/4c0014f5-89f0-48b6-88dc-cc4c2dba666a)
![Screenshot 2024-09-20 at 14 52
25](https://github.com/user-attachments/assets/202e5e1a-7c58-4af1-a85a-399c94313f0b)

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>