Move nested declarations above rule (#193931)
## Summary

> Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an
upcoming
version. To keep the existing behavior, move the declaration above the
nested
rule. To opt into the new behavior, wrap the declaration in & {}.

I have moved the declarations above the nested rules.

I had attempted to opt in to the new behavior, but the resulting CSS has
duplicate selectors, which CI does not allow

For more information about the new behavior, see
[here](https://sass-lang.com/documentation/breaking-changes/mixed-decls/)

Closes https://github.com/elastic/kibana/issues/190898
[screenshots](https://github.com/elastic/kibana/pull/193931#issuecomment-2375127180)
Closes https://github.com/elastic/kibana/issues/190899
[screenshots](https://github.com/elastic/kibana/pull/193931#issuecomment-2375128228)
Closes https://github.com/elastic/kibana/issues/190900 Access Page uses
`authentication_state_page`, but I was unable to find
'overwrite_session` and `logged_out` pages
Closes https://github.com/elastic/kibana/issues/190901
[screenshots](https://github.com/elastic/kibana/pull/193931#issuecomment-2375126897)
Closes https://github.com/elastic/kibana/issues/190902 ^used on same UI
Closes https://github.com/elastic/kibana/issues/190903
[screenshots](https://github.com/elastic/kibana/pull/193931#issuecomment-2375127629)
Closes https://github.com/elastic/kibana/issues/190904
[screenshots](https://github.com/elastic/kibana/pull/193931#issuecomment-2375127865)