[Reporting] Delete the task from Task Manager when deleting a report (#192417)
## Summary

Closes https://github.com/elastic/kibana/issues/191676

This PR ensures that when a user deletes a report, if there is a backing
Task Manager task that references the report job, that task is deleted
as well. With these changes, the report job will truly be "cancelled"
when a report is deleted.

The risk of not cancelling the Task Manager task when deleting a report
is evident in server logs and in task monitoring, when automated tests
are run on the reporting features.

### How to test
Primarily, the API Integration test for the Reporting Datastream
exercises the changes in this PR.

### Checklist

- [x] Ensure the API/Functional tests are using the delete report
endpoint rather than directly deleting reports from the ES store.
* **To prevent this PR from becoming very large, this assurance is only
needed for serverless tests, and for tests that do not wait for the
report job to be completed before deleting the report document.**
* Those constrictions narrow the changes down to just the test on the
Reporting Datastream
- [x] Handle scenario of report document deleted before task begins or
completes
- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit dfee92802d386a9d1d64e62bd363695a2b860b31)