[Inventory] Add icons for each Entity row (#194543)
## Summary

Closes https://github.com/elastic/kibana/issues/192523.

In this PR we are adding icons to each Entity row.

As a requirement, we need to use `agent.name` for Services and
`cloud.provider` for Hosts and Containers to determine which icon to
display.

|Before|After|
|-|-|
|![Screenshot 2024-10-01 at 12 22
47](https://github.com/user-attachments/assets/9b0dd3e7-9e63-450d-8438-590001985c60)|![Screenshot
2024-10-01 at 12 15
11](https://github.com/user-attachments/assets/a016b979-0526-4658-9cfe-d9ab38aef485)|

### Implementation details for `cloud.provider`

In some cases, `cloud.provider` may contain a list of values instead of
a single value. The team has agreed to always use the first element of
the list, as it is theoretically the provider with the most assigned
documents.

>[!WARNING]
This is still an open discussion.

This results in differences when comparing the Entities Inventory with
the Hosts list.

|Entities|Hosts|
|-|-|
|![Screenshot 2024-10-01 at 12 16
02](https://github.com/user-attachments/assets/2c19a2ac-3e9b-4699-a464-c74dccc58e2e)|![Screenshot
2024-10-01 at 12 16
31](https://github.com/user-attachments/assets/57a420cc-de8b-4880-97d0-aa331fe947b8)|

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Katerina <kate@kpatticha.com>
(cherry picked from commit 932614ab38f2f9d0021dd4d2ce0733427a448d40)