[Response Ops][Task Manager] Handle errors in `getCapacity` function during task polling (#194759)
## Summary

* Moves the `getCapacity` call during task polling within the try/catch
so any errors with this function will be caught and logged under the
`Failed to poll for work` message and polling continues
* During `mget` claim strategy, perform a final check after the
`bulkGet` to check for tasks with a null `startedAt` value. If any tasks
meet this condition, log some basic info and manually assign the
`startedAt`. This is a stop-gap measure to ensure we understand why we
might be seeing tasks with null `startedAt` values. In the future we may
choose to filter out these tasks from running in this cycle.