es3: use global empty state when available
Updated serverless search to use the global empty state as the home when
it is enabled. It will also move the current home to a dedicated getting
started page.

When we enable the global empty start FF config by default in es3 we can
remove the conditionals added with this code.