[UII] Fix first time integration flow when using multiple spaces (#194350)
## Summary

Resolves #182736. This PR fixes the first time Elastic Agent user
onboarding flow (only shown on Cloud) by prefixing the user's current
space ID to the default agent policy that is created behind the scenes
if the user has opted into using Fleet's space awareness.

This makes it so that Fleet correctly identifies whether there are any
agents enrolled in the current space, and directs the user to the
multi-step onboarding flow for adding integrations if there are no
agents enrolled in the current space's scoped agent policies.

An example policy ID created this way may look like
`second-space:fleet-first-agent-policy`. If the current space is the
`default` space, the ID remains as `fleet-first-agent-policy`.

Due to prefixing using a `:`, I also had to put policy ID in quotes
where ever it appears in kueries for filtering, i.e. `policy_id:"<some
id>"`. This change comprises most of the changed files in this PR.