[Security][Serverless] Removed override when the quality gate runs in QA (#192885)
## Summary

An override was used so far when a commit was provided no matter
who/what/where the tests were running.

With this change. the override will not be used when the quality gate
runs in QA. The reason behind this change is that we want to simulate
the customer's behavior, so we will be creating a project with whichever
version is currently deployed in QA.

---------

Co-authored-by: Domenico Andreoli <domenico.andreoli@elastic.co>
(cherry picked from commit 9ce6206e7e023db27e650a2398841c6a3222f1fb)