[Investigation app] add entities route and investigation Contextual Insight (#194432)
## Summary

Adds a route that can be used to fetch entities related to an
investigation.

The route fetches associated entities by service name, host name, or
container id. It then identifies the associated indices and datastreams.

The discovered entities are passed to the contextual insight to inform
the LLM.


![image](https://github.com/user-attachments/assets/855a8d68-b039-4557-ba23-5661cd961021)

This PR represents the first step in developing an AI-informed
hypothesis at the beginning of the investigation. Over time, further
insights will be provided to the LLM to deepen it's investigative
analysis and propose a more helpful root cause hypothesis.

### Testing

1. Create some APM data. I'm using the otel demo and triggering a
failure via the flagd service. Since this is in flux, you can reach out
to me about this workflow. However, you can also create APM data via
`synth-trace`.
2. Create an custom threshold rule that you expect to trigger an alert.
I created mine to using `http.response.status_code: 500 /
http.response.status_code : *` and set a low threshold base on the
amount of failures in my current test data. Be sure to also group the
alert by `service.name`
3. Wait for the alert to fire, then visit the alert details page and
start an investigation
4. notice the contextual insight. Expand it to see more information

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>