[Core Rendering Styles] Fix bug with kibanaFullBodyHeight calculation (#193798)
A variable for the page height was not being calculated correctly when
the Kibana chrome is invisible. This PR fixes the issue by adding `px`
as a unit to a variable used as input in the calculation.

Unblocks: https://github.com/elastic/kibana/pull/193647

## Testing
It will help to test within the
https://github.com/elastic/kibana/pull/193647 PR branch. The following
screenshots show the before and after effects of having this fix:

**before**

![ghjkgjghgkhjgjgkjg-1](https://github.com/user-attachments/assets/d684bc77-dcc0-40ef-9271-8ae198582eea)

**after**

![ghjkgjghgkhjgjgkjg-2](https://github.com/user-attachments/assets/5198cc43-b721-4ce8-92f0-0b2bac9f5eb5)

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
Co-authored-by: Marco Vettorello <marco.vettorello@elastic.co>
(cherry picked from commit 1df458d2d605f5a3281c3b18bda9f433d41a7c05)