[Auto Import] Ask LLM to map to non-reserved ECS fields (#195168)
## Release Note

Automatic Import does not ask the LLM to map the fields to the reserved
ECS fields anymore.

## Summary

Previously we have given the LLM the whole list of ECS fields, but later
failed the validation if the LLM's suggested a mapping into one of the
reserved fields (like `event.created`). With these changes, we hide the
reserved fields from the LLM when creating the prompt, so the likelihood
of this happening is reduced.

We test with the Teleport integration (see GitHub).
---------

Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
(cherry picked from commit c4599e0ade1aa259bc8ca62ed9a45af33d71816c)