[Obs AI Assistant] Avoid showing error toast when action is aborted b… (#195331)
Closes https://github.com/elastic/kibana/issues/186794

## Summary

### Problem
When the user clicks on "Stop generating" while the AI Assistant is
generating an answer, an error toast is shown with the error message
"Aborted". This is unnecessary as the user knows they aborted their
action.

### Solution
Do not show an error toast if the action is aborted by the user.

(cherry picked from commit 43a3d4a68ada67a584686b223ea8d798f4018cde)