[Synthetics] Fix ping heatmap payload (#195107)
## Summary

We addressed https://github.com/elastic/kibana/issues/180076 recently
with these two PRs:

- https://github.com/elastic/kibana/pull/184177
- https://github.com/elastic/kibana/pull/192508

We were seeing a strange error that was difficult to repro, so we put in
a best-effort patch that was still ineffective. The reason this issue
happens is because in the code it's possible to divide by 0, which
yields a value of `Infinity`, which at some point causes our interval
value supplied to the server route to be an empty string.

This patch will make it so that we never pass a value of 0 to be used in
the calculation of bucket sizes in this hook.

(cherry picked from commit 560d561e21fe51020954bd9e3246f238ffa026ba)