[Lens] Fix partition theme after ech upgrade (#195269)
## Summary

When I've updated elastic-charts with the last new version, the
`point.fill` parameter of lineSeriesStyle has changed in value. This
created an unwanted change also in the partition chart that was using
that style to color the partition sector borders.

I've removed the useless color override of `sectorLineStroke` in the
overwrite partition theme, leaving the control to use the chart theme.

I've also removed the possibility of unwanted changes for other
properties like `fontFamily` and in the `linkedText.textColor`.

In the same PR I've removed duplicated tests that where testing exactly
the same code/arguments/data, the only difference was the test name.

I've also refactored a bit the code, cleaning up the typings and
consolidating a bit the theme override logic

(cherry picked from commit 653073f8c6d199b069830c3acc13fb0c6664e9d8)